requirements.txt
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
colorzero==2.0
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-pubsub==2.4.1
googleapis-common-protos==1.53.0
gpiozero==1.6.2
grpc-google-iam-v1==0.12.3
grpcio==1.37.1
idna==2.10
libcst==0.3.18
mypy-extensions==0.4.3
packaging==20.9
picamera==1.13
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
RPi.GPIO==0.7.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
